+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog ##
 #### dev ####
 * Add missing property for imgui flag
+* Fix crash on Steam Deck by forcing LWJGL3 mode
 
 #### v3.30.0 ####
 * Make `Loader.ARGS` public so mods can check program arguments